Clean up TelemetryContainerFilter
- Make the naming of the attribute getters consistent
- Refactor the caching of routes out into its own class
