+++ dev/io.openliberty.microprofile.telemetry.1.0.internal/src/io/openliberty/microprofile/telemetry/internal/rest/TelemetryContainerFilter.java
@@ -4,7 +4,7 @@
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-2.0/
- * 
+ *
  * SPDX-License-Identifier: EPL-2.0
  *
  * Contributors:
@@ -15,13 +15,9 @@ package io.openliberty.microprofile.telemetry.internal.rest;
 import static java.util.Collections.emptyList;
 import static java.util.Collections.singletonList;
 
-import java.lang.ref.ReferenceQueue;
-import java.lang.ref.WeakReference;
 import java.lang.reflect.Method;
 import java.net.URI;
 import java.util.List;
-import java.util.Objects;
-import java.util.concurrent.ConcurrentHashMap;
 
 import io.openliberty.microprofile.telemetry.internal.helper.AgentDetection;
 import io.opentelemetry.api.OpenTelemetry;
@@ -60,122 +56,15 @@ public class TelemetryContainerFilter implements ContainerRequestFilter, Contain
     private static final String SPAN_PARENT_CONTEXT = "otel.span.server.parentContext";
     private static final String SPAN_SCOPE = "otel.span.server.scope";
 
-    private static final ServerAttributesExtractor serverAttributesExtractor = new ServerAttributesExtractor();
-    private static final NetServerAttributesGetterImpl netServerAttributesGetter = new NetServerAttributesGetterImpl();
+    private static final HttpServerAttributesGetterImpl HTTP_SERVER_ATTRIBUTES_GETTER = new HttpServerAttributesGetterImpl();
+    private static final NetServerAttributesGetterImpl NET_SERVER_ATTRIBUTES_GETTER = new NetServerAttributesGetterImpl();
 
-    private static final ConcurrentHashMap<RestRouteKey, String> routes = new ConcurrentHashMap<>();
-
-    private static final ReferenceQueue<Class<?>> referenceQueue = new ReferenceQueue<>();
-
-    @SuppressWarnings("unchecked")
-    private static void poll() {
-        RestRouteKeyWeakReference<Class<?>> key;
-        while ((key = (RestRouteKeyWeakReference<Class<?>>) referenceQueue.poll()) != null) {
-            routes.remove(key.getOwningKey());
-        }
-    }
-
-    private static String getRoute(Class<?> restClass, Method restMethod) {
-        poll();
-        return routes.get(new RestRouteKey(restClass, restMethod));
-    }
-
-    /**
-     * Add a new route for the specified REST Class and Method.
-     *
-     * @param restClass
-     * @param restMethod
-     * @param route
-     */
-    private static void putRoute(Class<?> restClass, Method restMethod, String route) {
-        poll();
-        routes.put(new RestRouteKey(referenceQueue, restClass, restMethod), route);
-    }
-
-    private static class RestRouteKey {
-        private final RestRouteKeyWeakReference<Class<?>> restClassRef;
-        private final RestRouteKeyWeakReference<Method> restMethodRef;
-        private final int hash;
-
-        RestRouteKey(Class<?> restClass, Method restMethod) {
-            this.restClassRef = new RestRouteKeyWeakReference<>(restClass, this);
-            this.restMethodRef = new RestRouteKeyWeakReference<>(restMethod, this);
-            hash = Objects.hash(restClass, restMethod);
-        }
-
-        RestRouteKey(ReferenceQueue<Class<?>> referenceQueue, Class<?> restClass, Method restMethod) {
-            this.restClassRef = new RestRouteKeyWeakReference<>(restClass, this, referenceQueue);
-            this.restMethodRef = new RestRouteKeyWeakReference<>(restMethod, this);
-            hash = Objects.hash(restClass, restMethod);
-        }
-
-        @Override
-        public int hashCode() {
-            return hash;
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (this == obj)
-                return true;
-            if (obj == null)
-                return false;
-            if (getClass() != obj.getClass())
-                return false;
-            RestRouteKey other = (RestRouteKey) obj;
-            if (!restClassRef.equals(other.restClassRef)) {
-                return false;
-            }
-            if (!restMethodRef.equals(other.restMethodRef)) {
-                return false;
-            }
-            return true;
-        }
-    }
-
-    private static class RestRouteKeyWeakReference<T> extends WeakReference<T> {
-        private final RestRouteKey owningKey;
-
-        RestRouteKeyWeakReference(T referent, RestRouteKey owningKey) {
-            super(referent);
-            this.owningKey = owningKey;
-        }
-
-        RestRouteKeyWeakReference(T referent, RestRouteKey owningKey,
-                                  ReferenceQueue<T> referenceQueue) {
-            super(referent, referenceQueue);
-            this.owningKey = owningKey;
-        }
-
-        RestRouteKey getOwningKey() {
-            return owningKey;
-        }
-
-        @SuppressWarnings("rawtypes")
-        @Override
-        public boolean equals(Object obj) {
-            if (obj == this) {
-                return true;
-            }
-
-            if (obj instanceof RestRouteKeyWeakReference) {
-                return get() == ((RestRouteKeyWeakReference) obj).get();
-            }
-
-            return false;
-        }
-
-        @Override
-        public String toString() {
-            T referent = get();
-            return new StringBuilder("RestRouteKeyWeakReference: ").append(referent).toString();
-        }
-    }
+    private static final RestRouteCache ROUTE_CACHE = new RestRouteCache();
 
     private Instrumenter<ContainerRequestContext, ContainerResponseContext> instrumenter;
 
     @jakarta.ws.rs.core.Context
-    ResourceInfo resourceInfo;
+    private ResourceInfo resourceInfo;
 
     // RESTEasy requires no-arg constructor for CDI injection: https://issues.redhat.com/browse/RESTEASY-1538
     public TelemetryContainerFilter() {
@@ -187,12 +76,12 @@ public class TelemetryContainerFilter implements ContainerRequestFilter, Contain
         InstrumenterBuilder<ContainerRequestContext, ContainerResponseContext> builder = Instrumenter.builder(
                                                                                                               openTelemetry,
                                                                                                               INSTRUMENTATION_NAME,
-                                                                                                              HttpSpanNameExtractor.create(serverAttributesExtractor));
+                                                                                                              HttpSpanNameExtractor.create(HTTP_SERVER_ATTRIBUTES_GETTER));
 
         this.instrumenter = builder
-                        .setSpanStatusExtractor(HttpSpanStatusExtractor.create(serverAttributesExtractor))
-                        .addAttributesExtractor(HttpServerAttributesExtractor.create(serverAttributesExtractor))
-                        .addAttributesExtractor(NetServerAttributesExtractor.create(netServerAttributesGetter))
+                        .setSpanStatusExtractor(HttpSpanStatusExtractor.create(HTTP_SERVER_ATTRIBUTES_GETTER))
+                        .addAttributesExtractor(HttpServerAttributesExtractor.create(HTTP_SERVER_ATTRIBUTES_GETTER))
+                        .addAttributesExtractor(NetServerAttributesExtractor.create(NET_SERVER_ATTRIBUTES_GETTER))
                         .buildServerInstrumenter(new ContainerRequestContextTextMapGetter());
     }
 
@@ -268,7 +157,7 @@ public class TelemetryContainerFilter implements ContainerRequestFilter, Contain
 
     }
 
-    private static class ServerAttributesExtractor implements HttpServerAttributesGetter<ContainerRequestContext, ContainerResponseContext> {
+    private static class HttpServerAttributesGetterImpl implements HttpServerAttributesGetter<ContainerRequestContext, ContainerResponseContext> {
 
         @Override
         public String flavor(final ContainerRequestContext request) {
@@ -281,7 +170,7 @@ public class TelemetryContainerFilter implements ContainerRequestFilter, Contain
             Class<?> resourceClass = (Class<?>) request.getProperty(REST_RESOURCE_CLASS);
             Method resourceMethod = (Method) request.getProperty(REST_RESOURCE_METHOD);
 
-            String route = getRoute(resourceClass, resourceMethod);
+            String route = ROUTE_CACHE.getRoute(resourceClass, resourceMethod);
 
             if (route == null) {
 
@@ -295,7 +184,7 @@ public class TelemetryContainerFilter implements ContainerRequestFilter, Contain
                 }
 
                 route = template.toTemplate();
-                putRoute(resourceClass, resourceMethod, route);
+                ROUTE_CACHE.putRoute(resourceClass, resourceMethod, route);
             }
 
             return route;
